refactor(test): split test_bspline.py into three focused files
Split the 1667-line test_bspline.py by responsibility:
- test_bspline.py: init, properties, edge cases (~370 lines)
- test_bspline_evaluation.py: evaluation and derivatives (~320 lines)
- test_bspline_conversion.py: to_open, to_periodic, to/from Bezier (~580 lines)
